VRモードのUIリファクタリング。
